Refactor (split) plot tests
Split the plot tests into separate files, and extract shared helpers
